+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Bug Fixes
 
 -   Fix [#40](https://github.com/Mezryss/FVTT-Genesys/issues/40): Upgrading ranked talents was too restrictive.
+-   Fix [#42](https://github.com/Mezryss/FVTT-Genesys/issues/42): Journal entries for skill upgrades are showing the wrong rank.
+-   Fix [#44](https://github.com/Mezryss/FVTT-Genesys/issues/44): Ranked Talent effects don't stack per rank.
 
 <details>
 <summary>Previous Releases</summary>